feat: use `runtime-deps` image
The packages data installed with `pip install` is stored in `/home/1001/.local`. We can preserve this path in a volume to maintain it between containers.

It's important to note that this is ***NOT*** a safe practice for most containerize Python applications. The reason we are able to do this is because the `stable-diffusion-webui` project has implemented a dependency check before each startup.

- Change the image source for the service `stable-diffusion-webui-forge` in  `docker-compose.yml` and `docker-compose.forge.yml` to `runtime-deps` image
- Add new volume `local` for `stable-diffusion-webui` and `stable-diffusion-webui-forge` service
- Add the volume binding `local:/home/1001/.local`

Signed-off-by: 陳鈞 <jim60105@gmail.com>

build: switch Forge Neo to runtime deps image with debug

- Switch the Forge Neo compose service to use the `forge-neo-runtime-deps` image and update the container name accordingly
- Enable debug startup logging flags by default
- Turn on CUDA malloc and stream options for the Forge backend
- Add a new `local` volume alongside the existing `cache` and `repositories` volumes

Signed-off-by: Jim Chen <Jim@ChenJ.im>

chore: enable debugging and optimizations in docker-compose files

- Enable debugging log level and debug mode in both docker-compose.forge and docker-compose files
- Enable 'cuda-malloc' and 'cuda-stream' optimizations in docker-compose.forge file

Signed-off-by: 陳鈞 <jim60105@gmail.com>